listings/requirements.txt
asgiref==3.2.10
autopep8==1.5.4
certifi==2020.6.20
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.1.1
django-cors-headers==3.5.0
djangorestframework==3.11.1
djangorestframework-simplejwt==4.4.0
docker==4.3.1
drf-yasg==1.17.1
gunicorn==20.0.4
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.2
MarkupSafe==1.1.1
packaging==20.4
Pillow==7.2.0
pycodestyle==2.6.0
PyJWT==1.7.1
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
uritemplate==3.0.1
urllib3==1.25.10
websocket-client==0.57.0
